dont use segwit change addresses for b-cash/b-gold

+++ BRTransaction.c
@@ -125,6 +125,8 @@ void BRTxInputSetWitness(BRTxInput *input, const uint8_t *witness, size_t witLen
     }
 }
 
+// serializes a tx input for a signature pre-image
+// set input->amount to 0 to skip serializing the input amount in non-witness signatures
 static size_t _BRTxInputData(const BRTxInput *input, uint8_t *data, size_t dataLen)
 {
     size_t off = 0;
@@ -181,6 +183,8 @@ void BRTxOutputSetScript(BRTxOutput *output, const uint8_t *script, size_t scrip
     }
 }
 
+// serializes the tx output at index for a signature pre-image
+// an index of SIZE_MAX will serialize all tx outputs for SIGHASH_ALL signatures
 static size_t _BRTransactionOutputData(const BRTransaction *tx, uint8_t *data, size_t dataLen, size_t index)
 {
     BRTxOutput *output;
@@ -207,8 +211,8 @@ static size_t _BRTransactionWitnessData(const BRTransaction *tx, uint8_t *data, 
     BRTxInput input;
     int anyoneCanPay = (hashType & SIGHASH_ANYONECANPAY), sigHash = (hashType & 0x1f);
     size_t i, off = 0;
-    uint8_t scriptCode[] = { OP_DUP, OP_HASH160, 20, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                             OP_EQUALVERIFY, OP_CHECKSIG };
+    uint8_t scriptCode[] = { OP_DUP, OP_HASH160, 20, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                             0, 0, 0, 0, 0, 0, 0, 0, 0, OP_EQUALVERIFY, OP_CHECKSIG };
 
     if (index >= tx->inCount) return 0;
     if (data && off + sizeof(uint32_t) <= dataLen) UInt32SetLE(&data[off], tx->version); // tx version
@@ -422,7 +426,7 @@ BRTransaction *BRTransactionParse(const uint8_t *buf, size_t bufLen)
     if (tx->inCount == 0 && off + 1 <= bufLen) witnessFlag = buf[off++];
     
     if (witnessFlag) {
-        tx->inCount = BRVarInt(&buf[off], (off <= bufLen ? bufLen - off : 0), &len);
+        tx->inCount = (size_t)BRVarInt(&buf[off], (off <= bufLen ? bufLen - off : 0), &len);
         off += len;
     }
 
@@ -467,11 +471,11 @@ BRTransaction *BRTransactionParse(const uint8_t *buf, size_t bufLen)
     
     for (i = 0, witnessOff = off; witnessFlag && off <= bufLen && i < tx->inCount; i++) {
         input = &tx->inputs[i];
-        count = BRVarInt(&buf[off], (off <= bufLen ? bufLen - off : 0), &len);
+        count = (size_t)BRVarInt(&buf[off], (off <= bufLen ? bufLen - off : 0), &len);
         off += len;
         
         for (j = 0, sLen = 0; j < count; j++) {
-            sLen += BRVarInt(&buf[off + sLen], (off + sLen <= bufLen ? bufLen - (off + sLen) : 0), &len);
+            sLen += (size_t)BRVarInt(&buf[off + sLen], (off + sLen <= bufLen ? bufLen - (off + sLen) : 0), &len);
             sLen += len;
         }
